LPS-99657 Rename 'update' to 'upgrade' for consistency
