+++ modules/apps/asset/asset-categories-selector-web/src/main/java/com/liferay/asset/categories/selector/web/internal/upgrade/AssetCategoriesSelectorWebUpgrade.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.asset.categories.selector.web.internal.update;
+package com.liferay.asset.categories.selector.web.internal.upgrade;
 
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
